feat(rust-bridge): M6 type-model tail — integer boundary tags + HashMap→dict marshaling
Two additive ABI-v1 tags round out the scalar/collection return model:

M6.1 — TAG_INT / TAG_UINT integer boundary tags. Signedness now rides the
tag, so a Rust i64/u64 return marshals with the correct reinterpret on both
loaders (byte-identical na↔CPython, proven by scalar_conformance against a
single .so). Seed: jac-bridge-scalar.

M6.2 — TAG_MAP_BIT (0x2000_0000, OR'd with the value tag). A HashMap<String,V>
return (V in int/uint/str/bool, incl. Result-wrapped) marshals as a real Jac
dict[str, V] via a one-shot deep copy over an owned JacBuf wire image
([count] then [key_len][key][value] per entry), freed through
jac_<mod>_free_buf. The CPython loader implements it end to end; the na
backend cannot yet build or return a dict (no dynamic subscript-assign,
dict-return ICEs, no unpack_from), so its synthesizer skips those methods with
a clear reason rather than emit code that would crash nacompile — proven
still-compilable by map_na_skip.jac. Seed: jac-bridge-map.

Layers: schema constants; jac-bridge macro (Tag::Map + extract_hashmap +
out_param serializer + vec_to_jacbuf); _blob tag helpers; _ctypes_codegen
map/int decode; _synth na-gate. Tests: test_scalar/test_map CPython loaders,
test_na_codegen_* source-shape, scalar_conformance + map_na_skip on na, ABI
drift guard, trybuild rejections for non-String keys / unsupported values. CI
wired in rust-bridges.yml.

+++ bridges/jac-bridge/tests/ui/unsupported_param_type.rs
@@ -5,8 +5,8 @@ mod m {
     pub struct Foo(pub u8);
 
     impl Foo {
-        // i32 is not a bridgeable boundary type.
-        pub fn bad(&self, x: i32) -> bool {
+        // f64 is not a bridgeable boundary type (no float tag in the v1 ABI).
+        pub fn bad(&self, x: f64) -> bool {
             let _ = x;
             true
         }
